Admin calendar script and template seperated.
